+++ scrapper/src/main/java/edu/java/scrapper/repository/jooq/JooqLinkRepository.java
@@ -89,24 +89,24 @@ public class JooqLinkRepository implements LinkRepository {
         }
 
         dslContext.insertInto(CHAT_LINK, CHAT_LINK.CHAT_ID, CHAT_LINK.LINK_ID)
-            .values(chatId, savedLink.get().getId().intValue()).execute();
+            .values(chatId, savedLink.get().getId()).execute();
         return savedLink.get();
     }
 
     @Override
-    public void delete(long chatId, long linkId) {
+    public void delete(long chatId, int linkId) {
         dslContext.deleteFrom(CHAT_LINK)
             .where(CHAT_LINK.CHAT_ID.eq(chatId))
-            .and(CHAT_LINK.LINK_ID.eq((int) linkId))
+            .and(CHAT_LINK.LINK_ID.eq(linkId))
             .execute();
 
         List<Long> chatIdsWithThisLink = dslContext
             .select(CHAT_LINK.CHAT_ID).from(CHAT_LINK)
-            .where(CHAT_LINK.LINK_ID.eq((int) linkId))
+            .where(CHAT_LINK.LINK_ID.eq(linkId))
             .fetchInto(Long.class);
 
         if (chatIdsWithThisLink.isEmpty()) {
-            dslContext.deleteFrom(LINK).where(LINK.ID.eq((int) linkId)).execute();
+            dslContext.deleteFrom(LINK).where(LINK.ID.eq(linkId)).execute();
         }
     }
 
@@ -115,7 +115,7 @@ public class JooqLinkRepository implements LinkRepository {
         dslContext
             .update(LINK)
             .set(LINK.LAST_CHECK_TIME, lastCheckTime)
-            .where(LINK.ID.eq(link.getId().intValue()))
+            .where(LINK.ID.eq(link.getId()))
             .execute();
     }
 }
